pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "jarvis"
version = "0.1.0"
description = "J.A.R.V.I.S. — Just A Really Very Intelligent System"
requires-python = ">=3.11"
dependencies = [
    "openwakeword>=0.6.0",
    "faster-whisper>=1.0.0",
    "mlx-whisper>=0.4.3; platform_system == 'Darwin' and platform_machine == 'arm64'",
    "sounddevice>=0.4.6",
    "numpy>=1.24",
    "anthropic>=0.25.0",
    "openai>=1.12.0",
    "edge-tts>=6.1.0",
    "pyyaml>=6.0",
    "piper-tts>=1.4.0",
    "psutil>=5.9",
    "pynput>=1.7.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
jarvis = "jarvis.main:run"

[tool.setuptools.packages.find]
include = ["jarvis*"]
